requirements_CUDA.txt
# Core Python deps (same as requirements.txt)
numpy==1.24.4
scipy
scikit-learn
scikit-dimension
torchvision
tqdm
pandas

# GPU packages (install via conda to match cluster CUDA runtime)
# Note: Install PyTorch with the matching CUDA toolchain from PyTorch instructions.
# Example (on cluster with CUDA 12.7 follow official instructions at https://pytorch.org):
#   conda install pytorch torchvision torchaudio pytorch-cuda -c pytorch -c nvidia
# FAISS (recommended via conda-forge or pytorch channels). Pin a version compatible
# with `numpy==1.24.4` to avoid ABI mismatches. Examples:
# CPU (conda-forge):
#   conda install -c conda-forge faiss-cpu=1.7.3
# GPU (pytorch + nvidia channels):
#   conda install -c pytorch -c nvidia faiss-gpu=1.7.3
# Note: prefer conda installs for FAISS to ensure the binary is built against
# the same NumPy ABI as the rest of the environment. Do NOT install FAISS via
# pip on the cluster unless you know the wheel was built for NumPy 1.24.x.

# Do NOT put a torch wheel here — pick the conda command above that matches the CUDA
# version on the target machine (12.7). See README.md for more guidance.